implement(invoice): add ModifyInvoice method for line item modifications
